requirements.txt
colorzero==2.0
gpiozero==2.0.1
pygame==2.6.1
RPi.GPIO==0.7.1
